Create README.md
A Live Quote with Different Images this program allows users to see a motivational or inspirational quote along with a different image each time the quote is generated. Upon clicking a button, the application fetches a new quote and displays it alongside a random image from a predefined set of images.